attempting to export ts ext to other files
