+++ src/server/server.ts
@@ -1,25 +1,7 @@
 import express, { Request, Response, NextFunction } from 'express';
 // const path = require('path');
 import path from 'path';
-import { Pool } from 'pg';
-import dotenv from 'dotenv';
-
-dotenv.config(); // crucial
-
-const connectionString: string = process.env.CONNECTION_STRING
-const client = new Pool({
-  connectionString,
-});
-
-const connectDB = async () => {
-  try {
-    await client.connect();
-    console.log('Connected to ElephantSQL database');
-  } catch (error) {
-    console.error('Failed to connect to the database', error);
-  }
-};
-
+import connectDB from './database.ts';
 
 const app = express();
